Funcionando

+++ src/dao/DAOHostal.java
@@ -78,7 +78,7 @@ public class DAOHostal
 	 * @throws SQLException SQLException Genera excepcion si hay error en la conexion o en la consulta SQL
 	 * @throws Exception Si se genera un error dentro del metodo.
 	 */
-	public Hostal findHostalById(Long id) throws SQLException, Exception 
+	public Hostal findHostalById(Integer id) throws SQLException, Exception 
 	{
 		Hostal Hostal = null;
