+++ src/dao/DAOReserva.java
@@ -82,7 +82,7 @@ public class DAOReserva {
 	 * @throws SQLException SQLException Genera excepcion si hay error en la conexion o en la consulta SQL
 	 * @throws Exception Si se genera un error dentro del metodo.
 	 */
-	public Reserva findReservaById(Long id) throws SQLException, Exception 
+	public Reserva findReservaById(Integer id) throws SQLException, Exception 
 	{
 		Reserva Reserva = null;
