+++ src/dao/DAOVecino.java
@@ -80,7 +80,7 @@ public class DAOVecino {
 	 * @throws SQLException SQLException Genera excepcion si hay error en la conexion o en la consulta SQL
 	 * @throws Exception Si se genera un error dentro del metodo.
 	 */
-	public Vecino findVecinoById(Long id) throws SQLException, Exception 
+	public Vecino findVecinoById(Integer id) throws SQLException, Exception 
 	{
 		Vecino Vecino = null;
