+++ src/rest/ReservaService.java
@@ -87,7 +87,7 @@ public class ReservaService <K extends Operador>{
 	@GET
 	@Path( "{id: \\d+}" )
 	@Produces( { MediaType.APPLICATION_JSON } )
-	public Response getReservaById( @PathParam( "id" ) Long id )
+	public Response getReservaById( @PathParam( "id" ) Integer id )
 	{
 		try{
 			AlohAndesTransactionManager<K> tm = new AlohAndesTransactionManager<K>( getPath( ) );
